refactor: dedupe stdin cli providers
